lab_2/users/requirements.txt
fastapi
uvicorn
pydantic
email-validator
python-jose
passlib==1.7.4
python-multipart
bcrypt==4.0.1